requirements.txt
numpy==1.24.3
opencv_python==4.7.0.72
pandas==1.5.3
Pillow==9.4.0
Pillow==9.5.0
streamlit==1.23.1
streamlit_option_menu==0.3.6
